Main logic partialy created
In this commit a lot of architecture was created.
Main packages were added as well as main classes that will be resposible
for this application logic. The MVC model is here implemented.
The DTO pattern for sending objets is applied.
The application logic will be deployd from this moment forward